ci, Docker compose and some more (#43)
The most part of the changes are about server: fixed some troubles with
MEDIA_DIR, now it is created automatically if not exists.

Added simple ci pipeline with two stages: linting with flake-8 and tests
with pytest, both for backend. This stages run in every branch.

Requirements for backend were split to base, local, docker, and test.

Local (aka development) configuration assumes to run project with SQLite
database and standard Python dict for user sessions.
Docker (aka production) configuration assumes to run project with Docker
Compose, Postgres as a database and a Redis for user sessions.